Share the mapper inset test helpers
The inset editor tests are about to grow a second file that needs the same
map and handle selectors, so they move out of the shard-running closure in
mapper-edit-insets.ts and into mapper-utils.ts. The wheel event helper goes
to test_utils.ts, next to the other synthetic event helpers.
